fix arm64 build err

Created-by: liubowen26188
Commit-by: liubowen26188
Merged-by: ohci1
Description: ### 相关的Issue
[#6540](https://gitcode.com/openharmony/drivers_peripheral/issues/6540) 

### 原因（目的、解决的问题等）
解决arm64编译错误

### 描述（做了什么，变更了什么）
兼容arm32 arm64编译，强制类型转换

### 测试用例（新增、改动、可能影响的功能）






See merge request: openharmony/drivers_peripheral!8590

+++ usb/hdi_service/src/usb_device_impl.cpp
@@ -453,7 +453,7 @@ std::string UsbDeviceImpl::GetDeviceDirName(uint8_t busNum, uint8_t devAddr)
         }
         if (busNumConverted > UINT8_MAX || devAddrConverted > UINT8_MAX) {
             HDF_LOGE("%{public}s: value out of range: busNum=%{public}llu, devAddr=%{public}llu",
-                __func__, busNumConverted, devAddrConverted);
+                __func__, (unsigned long long)busNumConverted, (unsigned long long)devAddrConverted);
             deviceDir = "";
             continue;
         }
